# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gestion des Réclamations Étudiant",
    "description": "API REST pour la gestion des réclamations et tickets étudiants avec authentification JWT et contrôle d'accès basé sur les rôles.",
    "version": "1.0.0",
    "contact": {
      "name": "Support API",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Serveur de développement"
    },
    {
      "url": "https://api.example.com",
      "description": "Serveur de production"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Connexion utilisateur",
        "description": "Authentifie un utilisateur avec email et mot de passe, retourne un token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "admin@example.com",
                "password": "Admin123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Déconnexion utilisateur",
        "description": "Invalide le token de rafraîchissement de l'utilisateur",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "description": "Token de rafraîchissement à invalider"
                  }
                },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Déconnexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Utilisateurs"],
        "summary": "Lister les utilisateurs",
        "description": "Récupère la liste paginée des utilisateurs avec filtres optionnels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre d'éléments par page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Filtrer par rôle",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ADMIN", "QA", "STO", "VIEWER"]
            }
          },
          {
            "name": "department",
            "in": "query",
            "description": "Filtrer par département",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Utilisateurs"],
        "summary": "Créer un utilisateur",
        "description": "Crée un nouvel utilisateur (ADMIN uniquement)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Email déjà utilisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Mot de passe de l'utilisateur"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string",
                "description": "Token JWT d'accès"
              },
              "refresh_token": {
                "type": "string",
                "description": "Token JWT de rafraîchissement"
              },
              "user": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID unique de l'utilisateur"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email de l'utilisateur"
          },
          "nom": {
            "type": "string",
            "description": "Nom de famille"
          },
          "prenom": {
            "type": "string",
            "description": "Prénom"
          },
          "telephone": {
            "type": "string",
            "description": "Numéro de téléphone"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "SUSPENDED"],
            "description": "Statut de l'utilisateur"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ADMIN", "QA", "STO", "VIEWER"]
            },
            "description": "Rôles de l'utilisateur"
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Départements de l'utilisateur"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de dernière modification"
          }
        }
      },
      "UserCreate": {
        "type": "object",
        "required": ["email", "password", "nom", "prenom"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email de l'utilisateur"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Mot de passe de l'utilisateur"
          },
          "nom": {
            "type": "string",
            "description": "Nom de famille"
          },
          "prenom": {
            "type": "string",
            "description": "Prénom"
          },
          "telephone": {
            "type": "string",
            "description": "Numéro de téléphone"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ADMIN", "QA", "STO", "VIEWER"]
            },
            "description": "Rôles à assigner"
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Départements à assigner"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "PaginatedUsersResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Page actuelle"
          },
          "limit": {
            "type": "integer",
            "description": "Nombre d'éléments par page"
          },
          "total": {
            "type": "integer",
            "description": "Nombre total d'éléments"
          },
          "pages": {
            "type": "integer",
            "description": "Nombre total de pages"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Code d'erreur"
              },
              "message": {
                "type": "string",
                "description": "Message d'erreur"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Détails de l'erreur"
              }
            }
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Message de succès"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentification",
      "description": "Endpoints d'authentification et de gestion des sessions"
    },
    {
      "name": "Utilisateurs",
      "description": "Gestion des utilisateurs, rôles et départements"
    },
    {
      "name": "Tickets",
      "description": "Gestion des tickets et réclamations"
    },
    {
      "name": "Commentaires",
      "description": "Gestion des commentaires sur les tickets"
    },
    {
      "name": "Analytics",
      "description": "Statistiques et rapports"
    }
  ]
}